api/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pip==20.0.2
pkg-resources==0.0.0
PyJWT==2.4.0
pytz==2022.1
setuptools==44.0.0
sqlparse==0.4.2
